Placate clippy

+++ crates/services/executor/src/executor.rs
@@ -191,7 +191,7 @@ impl TransactionsSource for OnceTransactionsSource {
         let transactions: &mut Vec<MaybeCheckedTransaction> = lock.as_mut();
         // Avoid panicking if we request more transactions than there are in the vector
         let transactions_limit = (transactions_limit as usize).min(transactions.len());
-        transactions.drain(..transactions_limit as usize).collect()
+        transactions.drain(..transactions_limit).collect()
     }
 }
